Moved UtilTests from test_urls.py to test_util.py
